Move cli tests to the cli subpackage
